+++ sdk/resourcemanager/networkfunction/armnetworkfunction/azuretrafficcollectors_client.go
@@ -46,7 +46,7 @@ func NewAzureTrafficCollectorsClient(subscriptionID string, credential azcore.To
 // BeginCreateOrUpdate - Creates or updates a Azure Traffic Collector resource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - parameters - The parameters to provide for the created Azure Traffic Collector.
@@ -73,7 +73,7 @@ func (client *AzureTrafficCollectorsClient) BeginCreateOrUpdate(ctx context.Cont
 // CreateOrUpdate - Creates or updates a Azure Traffic Collector resource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 func (client *AzureTrafficCollectorsClient) createOrUpdate(ctx context.Context, resourceGroupName string, azureTrafficCollectorName string, parameters AzureTrafficCollector, options *AzureTrafficCollectorsClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AzureTrafficCollectorsClient.BeginCreateOrUpdate"
@@ -115,7 +115,7 @@ func (client *AzureTrafficCollectorsClient) createOrUpdateCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -127,7 +127,7 @@ func (client *AzureTrafficCollectorsClient) createOrUpdateCreateRequest(ctx cont
 // BeginDelete - Deletes a specified Azure Traffic Collector resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - options - AzureTrafficCollectorsClientBeginDeleteOptions contains the optional parameters for the AzureTrafficCollectorsClient.BeginDelete
@@ -153,7 +153,7 @@ func (client *AzureTrafficCollectorsClient) BeginDelete(ctx context.Context, res
 // Delete - Deletes a specified Azure Traffic Collector resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 func (client *AzureTrafficCollectorsClient) deleteOperation(ctx context.Context, resourceGroupName string, azureTrafficCollectorName string, options *AzureTrafficCollectorsClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AzureTrafficCollectorsClient.BeginDelete"
@@ -195,7 +195,7 @@ func (client *AzureTrafficCollectorsClient) deleteCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -204,7 +204,7 @@ func (client *AzureTrafficCollectorsClient) deleteCreateRequest(ctx context.Cont
 // Get - Gets the specified Azure Traffic Collector in a specified resource group
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - options - AzureTrafficCollectorsClientGetOptions contains the optional parameters for the AzureTrafficCollectorsClient.Get
@@ -251,7 +251,7 @@ func (client *AzureTrafficCollectorsClient) getCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -269,7 +269,7 @@ func (client *AzureTrafficCollectorsClient) getHandleResponse(resp *http.Respons
 // UpdateTags - Updates the specified Azure Traffic Collector tags.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - parameters - Parameters supplied to update Azure Traffic Collector tags.
@@ -317,7 +317,7 @@ func (client *AzureTrafficCollectorsClient) updateTagsCreateRequest(ctx context.
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {