+++ sdk/resourcemanager/networkfunction/armnetworkfunction/azuretrafficcollectorsbyresourcegroup_client.go
@@ -45,7 +45,7 @@ func NewAzureTrafficCollectorsByResourceGroupClient(subscriptionID string, crede
 
 // NewListPager - Return list of Azure Traffic Collectors in a Resource Group
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - options - AzureTrafficCollectorsByResourceGroupClientListOptions contains the optional parameters for the AzureTrafficCollectorsByResourceGroupClient.NewListPager
 //     method.
@@ -88,7 +88,7 @@ func (client *AzureTrafficCollectorsByResourceGroupClient) listCreateRequest(ctx
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil