+++ sdk/resourcemanager/networkfunction/armnetworkfunction/collectorpolicies_client.go
@@ -46,7 +46,7 @@ func NewCollectorPoliciesClient(subscriptionID string, credential azcore.TokenCr
 // BeginCreateOrUpdate - Creates or updates a Collector Policy resource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - collectorPolicyName - Collector Policy Name
@@ -74,7 +74,7 @@ func (client *CollectorPoliciesClient) BeginCreateOrUpdate(ctx context.Context, 
 // CreateOrUpdate - Creates or updates a Collector Policy resource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 func (client *CollectorPoliciesClient) createOrUpdate(ctx context.Context, resourceGroupName string, azureTrafficCollectorName string, collectorPolicyName string, parameters CollectorPolicy, options *CollectorPoliciesClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "CollectorPoliciesClient.BeginCreateOrUpdate"
@@ -120,7 +120,7 @@ func (client *CollectorPoliciesClient) createOrUpdateCreateRequest(ctx context.C
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -132,7 +132,7 @@ func (client *CollectorPoliciesClient) createOrUpdateCreateRequest(ctx context.C
 // BeginDelete - Deletes a specified Collector Policy resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - collectorPolicyName - Collector Policy Name
@@ -159,7 +159,7 @@ func (client *CollectorPoliciesClient) BeginDelete(ctx context.Context, resource
 // Delete - Deletes a specified Collector Policy resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 func (client *CollectorPoliciesClient) deleteOperation(ctx context.Context, resourceGroupName string, azureTrafficCollectorName string, collectorPolicyName string, options *CollectorPoliciesClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "CollectorPoliciesClient.BeginDelete"
@@ -205,7 +205,7 @@ func (client *CollectorPoliciesClient) deleteCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -214,7 +214,7 @@ func (client *CollectorPoliciesClient) deleteCreateRequest(ctx context.Context, 
 // Get - Gets the collector policy in a specified Traffic Collector
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - collectorPolicyName - Collector Policy Name
@@ -265,7 +265,7 @@ func (client *CollectorPoliciesClient) getCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -282,7 +282,7 @@ func (client *CollectorPoliciesClient) getHandleResponse(resp *http.Response) (C
 
 // NewListPager - Return list of Collector policies in a Azure Traffic Collector
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - options - CollectorPoliciesClientListOptions contains the optional parameters for the CollectorPoliciesClient.NewListPager
@@ -330,7 +330,7 @@ func (client *CollectorPoliciesClient) listCreateRequest(ctx context.Context, re
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -348,7 +348,7 @@ func (client *CollectorPoliciesClient) listHandleResponse(resp *http.Response) (
 // UpdateTags - Updates the specified Collector Policy tags.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-11-01
+// Generated from API version 2024-12-01
 //   - resourceGroupName - The name of the resource group.
 //   - azureTrafficCollectorName - Azure Traffic Collector name
 //   - collectorPolicyName - Collector Policy Name
@@ -401,7 +401,7 @@ func (client *CollectorPoliciesClient) updateTagsCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-11-01")
+	reqQP.Set("api-version", "2024-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {